refactor(core): extract state I/O and typed error classes
Move the state-management block (ensureDir, readJsonFile, writeJsonFile,
appendJsonl, readJsonl, statePaths, setPrivateFilePermissions, plus the
internal ensureParent helper) from skills/alien-agent-id/lib.mjs into
plugins/agent-id-core/lib/state.mjs. The state-dir layout — owner-session,
pending-auth, nonces, sequence, keys/, delegations/, audit/, vault/ — is
now defined in one place that every plugin can import.

Move SubjectMismatchError and AuthRevokedError into
plugins/agent-id-core/lib/errors.mjs. These are typed error classes
callers branch on via instanceof; centralizing them keeps the contract
stable as the lib splits across plugins.

lib.mjs re-exports both modules with `export *` and imports them for
internal use. No public-surface changes; all 184 tests pass.
